+++ mohupy/core/__init__.py
@@ -1,35 +1,33 @@
-#  Copyright (c) yibocat 2023 All Rights Reserved
+#  Copyright (c) yibocat 2024 All Rights Reserved
 #  Python: 3.10.9
-#  Date: 2023/11/26 下午2:47
+#  Date: 2024/4/6 上午11:47
 #  Author: yibow
 #  Email: yibocat@yeah.net
 #  Software: MohuPy
+
 __all__ = []
 
-from .nums import Fuzznum
-from .array import Fuzzarray
 from .base import FuzzType
+from .fuzzarray import Fuzzarray
+from .fuzznums import Fuzznum
 
 from .function import normalize, broadcast_to
 from .regedit import Registry
-from .package import *
+from .operationpackage import *
+from .operationLib import archimedeanDict
 
-__all__ += ['FuzzType',
-            'Fuzznum',
-            'Fuzzarray',
-            'normalize',
-            'Registry',
-            'broadcast_to']
+__all__ += ['FuzzType', 'Fuzzarray', 'Fuzznum',
+            'normalize', 'broadcast_to', 'Registry', 'archimedeanDict']
 
-from .function import (isValid, isEmpty, isInitial, convert, qsort, unique,
+from .function import (valid, empty, initial, convert, qsort, unique,
                        append, remove, pop, reshape, squeeze, clear,
-                       ravel, flatten, getmax, getmin, getsum, getprod, mean,
-                       fmax, fmin)
+                       ravel, flatten, getmax, getmin, getsum,getprod,
+                       mean, fmax, fmin)
 from .operation import (add, sub, mul, div, pow, matmul, equal, inequal,
                         lt, gt, le, ge, getitem)
 
 __all__ += [
-    'isValid', 'isEmpty', 'isInitial', 'convert', 'qsort', 'unique',
+    'valid', 'empty', 'initial', 'convert', 'qsort', 'unique',
     'append', 'remove', 'pop', 'reshape', 'squeeze', 'clear',
     'ravel', 'flatten', 'getmax', 'getmin', 'getsum', 'getprod', 'mean',
     'fmax', 'fmin',
@@ -37,6 +35,5 @@ __all__ += [
     'lt', 'gt', 'le', 'ge', 'getitem'
 ]
 
-from .fuzzfunc import fuzznum, fuzzset
-
+from .construct import fuzznum, fuzzset
 __all__ += ['fuzznum', 'fuzzset']
