+++ mohupy/core/attribute.py
@@ -1,0 +1,35 @@
+#  Copyright (c) yibocat 2024 All Rights Reserved
+#  Python: 3.10.9
+#  Date: 2024/4/6 下午2:06
+#  Author: yibow
+#  Email: yibocat@yeah.net
+#  Software: MohuPy
+from typing import Union
+
+from .attributeClass import *
+from .fuzzarray import Fuzzarray
+from .fuzznums import Fuzznum
+
+
+def report(x):
+    return Report()(x)
+
+
+def string(x):
+    return Str()(x)
+
+
+def score(x):
+    return Score()(x)
+
+
+def acc(x):
+    return Accuracy()(x)
+
+
+def ind(x):
+    return Indeterminacy()(x)
+
+
+def comp(x) -> Union[Fuzznum, Fuzzarray]:
+    return Complement()(x)
