+++ mohupy/core/regedit.py
@@ -1,9 +1,10 @@
-#  Copyright (c) yibocat 2023 All Rights Reserved
+#  Copyright (c) yibocat 2024 All Rights Reserved
 #  Python: 3.10.9
-#  Date: 2023/11/27 下午7:01
+#  Date: 2024/4/6 下午2:35
 #  Author: yibow
 #  Email: yibocat@yeah.net
 #  Software: MohuPy
+
 from collections.abc import MutableMapping
