+++ mohupy/tensor/__init__.py
@@ -1,6 +1,6 @@
-#  Copyright (c) yibocat 2023 All Rights Reserved
+#  Copyright (c) yibocat 2024 All Rights Reserved
 #  Python: 3.10.9
-#  Date: 2023/11/29 下午1:43
+#  Date: 2024/4/6 下午1:45
 #  Author: yibow
 #  Email: yibocat@yeah.net
 #  Software: MohuPy
