--- test/test.csv
@@ -1,4 +1,0 @@
-,C1,C2,C3,C4,C5
-A1,"<[0.4752 0.4183], [0.6126 0.0608]>","<[0.1558 0.6167 0.5553 0.6276], [0.4939 0.5197 0.6193]>","<[0.5108 0.1508], [0.7401 0.7684]>","<[0.5073 0.8168 0.688 ], [0.1219 0.5267 0.4272]>","<[0.4168 0.7811 0.3702 0.9724], [0.1933]>"
-A2,"<[0.6648 0.4696 0.0069 0.4488], [0.3628 0.834 ]>","<[0.2453], [0.2895 0.9885]>","<[0.2356 0.9849], [0.3083]>","<[0.853  0.1468 0.6939], [0.6329 0.5758]>","<[0.1815], [0.502  0.7817 0.8436]>"
-A3,"<[0.5844 0.458 ], [0.4332 0.1143 0.0988]>","<[0.3379], [0.8372 0.944 ]>","<[0.1404 0.2639], [0.7845]>","<[0.7858 0.6092], [0.3231 0.3694]>","<[0.0234], [0.3599 0.8197]>"
