--- test/test.py
@@ -1,11 +1,0 @@
-#  Copyright (c) yibocat 2023 All Rights Reserved
-#  Python: 3.10.9
-#  Date: 2023/12/19 下午8:53
-#  Author: yibow
-#  Email: yibocat@yeah.net
-#  Software: MohuPy
-
-import mohupy as mp
-
-x = mp.random.rand(3, 'qrofn', 3, 4)
-
